requirements.txt
Flask==0.12.2
google-cloud-storage==1.2.0